services/user_service/requirements.txt
# --- Web framework ---
fastapi==0.110.1
uvicorn[standard]==0.27.1

# --- Database ORM ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# --- AWS Integration ---
boto3==1.34.89

# --- HTTP client ---
requests==2.31.0
httpx==0.27.0

# --- Auth / Cognito (JWT) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Pydantic for validation ---
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# --- Optional (for local dev/testing) ---
python-dotenv==1.0.1

# --- For packaging or deployment (optional) ---
mangum==0.17.0
